# Domain records list performance improvement - removing Setting.get query from loop