I moved pretty_ipv6_ptr setting retrieval out of record list loop. It's massive waste of resources when having large zones leading to unacceptable response time in UI.
example.com zone with 4000 records
Request URL: /domain/example.com

4000 queries hit SQL server: "SELECT setting.id AS setting_id, setting.name AS setting_name, setting.value AS setting_value FROM setting WHERE setting.name = 'pretty_ipv6_ptr'  LIMIT 1"

Browser "network" inspection:
19.49s (16.77s waiting for server response)
Same request and inspection report after this commit:
4.83s (2.54s waiting for server response).